requirements.txt
discord.py>=2.0.0
python-dotenv
asyncpg
google-generativeai
rich
PyNaCl
psycopg2-binary
markdown
discord.py[voice]
uvloop